Add deployment code for a bastion Heat stack
Under tasks/main.yml add call to os_stack to create a Heat stack that
either has a random name generated by the playbook or a name selected by
the user. Add the Heat template for the stack under files/. This initial
version creates one virtual machine with an attached floating IP and a
security group to manage access.