Explicitly specify Heat template location
It looks like os_stack will look for the Heat template relative to the
playbook and not look under the files/ directory of the role as is the
case usually. To avoid this behaviour, explicitly define that the
template should be looked for under the files/ directory of the role.