Add cloud-config to the bastion server in Heat (#4)
It's useful to be able to customize the bastion host. Providing a
mechanism for the user to do this customization via cloud-config seems
like the best option. Add a new parameter to the Heat stack for passing
on cloud-init configuration. Document how to use this parameter in the
example environment file for Heat.

Also fix a bug in random name generation failure handling. To check for
the existence of stderr_lines, one must use "is defined".