Add group in inventory based on heat_stack_name (#6)
Create a group in the inventory that is based on the Heat stack name.
This makes it possible to assign group variables based on the Heat stack
name so that it becomes easier to manage multiple bastion hosts via a
single control repository. Also add the bastion host to this group when
adding it to the inventory with add_host.